Add files via upload
In this lab we were asked to create a Dynamic matrice, fill it with some data, then create a function which takes a matrice,and two rows and switch the content of the rows with O(1) running time.  the function must get a universal pointer to a matrice.
the second Question was to create a function which takes a matrice, two columns and the number of rows and replace the content of the two columns, O(n).