Add files via upload
in this lab the first question is to create a function which takes values from the user, insert it into two linked lists, one for odd numbers
and one for even. the amout of values the user inputted needs to be returned as the two linked lists need to be changed by reference alone.

the second question asks us to sum the list, return the sum and also by reference return the min and max values of the list.

in the end we needed to free all the allocated memory which was allocated for the lists.