Add files via upload
In this lab we get a linked list of characters (assuming these are lower case english letters) and in the first function we need to reverse the list
which means, we get a list of a , b , c ,d and get d,c,b,a.

in the second function we need to insert a new node in the list in the correct position as if the list is sorted.
if we have a,b,c,d and we insert another c it should be a,b,c,c,d and not c,a,b,c,d or a,b,c,d,c.
if we do not have a sorted list to begin with, the function will still insert the letter in the "correct" position though it will not make the list sorted. 
e,d,c,a 
insert(b)
b,e,d,c,a since b is before e alphabetically.